Redone project files
Had to redo the project files a bit more, had a built that wasn't fully working
