+++ SFML_Tutorial/Scene01.txt
@@ -1,1 +1,1 @@
-1019.000000,749.000000
+985.000000,77.000000
