+++ SFML_Tutorial/UI_Button.h
@@ -1,0 +1,5 @@
+#pragma once
+class UI_Button
+{
+};
+
